zaliv111

+++ CryptoGridTradingBot/Services/GridTradingEngine.cs
@@ -13,7 +13,7 @@ namespace CryptoGridTradingBot.Services
 
         public List<(double Buy, double Sell)> CalculateGridLevels(double currentPrice, double atr, double rsi, double volume, int levels)
         {
-            float optimalSpacing = _mlModel.PredictOptimalGridSpacing((float)atr, (float)rsi, (float)volume);
+            float optimalSpacing = (float)_mlModel.PredictOptimalGridSpacing((float)atr, (float)rsi, (float)volume);
             var gridLevels = new List<(double Buy, double Sell)>();
 
             for (int i = 1; i <= levels; i++)
